refactor(format): extract diff computation owner
