+++ ci/proof.toml
@@ -721,6 +721,7 @@ coverage = true
 name = "format_core_outputs"
 kind = "rust"
 paths = [
+  "crates/tokmd-format/src/diff/**",
   "crates/tokmd-format/src/diff.rs",
   "crates/tokmd-format/src/export.rs",
   "crates/tokmd-format/src/lib.rs",
